Add recording of snarf runs, using an Activity model. This will record the date and time of the snarf, the number of logs processed and, if there is an error, store the details for review.
Expands the Downloads record to include the Amazon request id from
the download log entry. This allows the provision of a safety net
where snarfs can be re-run at will and duplicates in the database
will be prevented.

Separated the database initialization into a different module, which
gives some flexibility around testing.

Added first couple of unit tests for the snarfer, with coverage
measuring capability.

Sundry refactorings to cut code size in the Snarfer.